Update interfaces for KP4.7

+++ Client_Interfaces/java/exemple/Mission1.java
@@ -1,6 +1,6 @@
 package exemple;
 
-import exemple.ConstantList_KP_4_1;
+import exemple.ConstantList_KP_4_7;
 
 import pp.PP;
 import pp.Unit;
@@ -9,7 +9,7 @@ import java.awt.geom.Point2D;
 public class Mission1 {
 	public static void main (String[] args){
 		// definition de la coordonnee ciblee a atteindre
-		Point2D.Float p = new Point2D.Float (1983, 1279);
+		Point2D.Float p = new Point2D.Float (1980, 1676);
 		try {
 			// creer une interface avec le jeu
 			PP pp = new PP();
